Validation is successfully moved to the new content iteration interface. Divided it to multiple pieces based on the area of validation the mixins are doing to make the package easier to navigate. Segmentation logic has 2 tests failing, which are going to be fixed once I moved them too to this interface. resolves #223
